fix: Change type from int to float64.

+++ search.go
@@ -171,7 +171,7 @@ type CustomOrderSearchReturn struct {
 		} `json:"orderCustomer"`
 		Currency struct {
 			IsoCode                        string      `json:"isoCode"`
-			Factor                         int         `json:"factor"`
+			Factor                         float64     `json:"factor"`
 			Symbol                         string      `json:"symbol"`
 			ShortName                      string      `json:"shortName"`
 			Name                           string      `json:"name"`
